move check_services_in_yaml_configs to ValidateServicesInSetsCommand
